refactor(guard): split guard.py into package
- guard.py hit 1.1k lines mixing five domain axes in one file
- move code verbatim into guard/ submodules, public API re-exported
  from __init__ so tool imports stay unchanged; tests mirror the split
