+++ src/mcp_server_polarion/tools/_shared/guard/__init__.py
@@ -1,0 +1,51 @@
+"""Pre-write guards: Polarion persists unknown enum ids / custom-field keys as
+silent ghosts (HTTP 200, invisible to UI and Lucene), so each guard fetches the
+real options and raises before the write. Fail-closed — validation error blocks
+the write (auth → ``PermissionError``, else ``RuntimeError``); only a
+*successful* empty option set and a 404 defer to Polarion. Caching in
+:mod:`...tools._shared.cache`.
+"""
+
+from __future__ import annotations
+
+from mcp_server_polarion.tools._shared.guard.documents import (
+    guard_document_custom_fields,
+    guard_document_enums,
+)
+from mcp_server_polarion.tools._shared.guard.enums import (
+    fetch_enum_option_ids,
+    fetch_project_enum_option_ids,
+)
+from mcp_server_polarion.tools._shared.guard.links import (
+    guard_hyperlink_roles,
+    guard_work_item_link_roles,
+    guard_work_item_link_targets,
+    partition_delete_links,
+)
+from mcp_server_polarion.tools._shared.guard.test_runs import (
+    guard_test_run_custom_fields,
+    guard_test_run_enums,
+    guard_test_run_templates,
+)
+from mcp_server_polarion.tools._shared.guard.work_items import (
+    guard_work_item_custom_fields,
+    guard_work_item_enums,
+    resolve_work_item_types,
+)
+
+__all__ = [
+    "fetch_enum_option_ids",
+    "fetch_project_enum_option_ids",
+    "guard_document_custom_fields",
+    "guard_document_enums",
+    "guard_hyperlink_roles",
+    "guard_test_run_custom_fields",
+    "guard_test_run_enums",
+    "guard_test_run_templates",
+    "guard_work_item_custom_fields",
+    "guard_work_item_enums",
+    "guard_work_item_link_roles",
+    "guard_work_item_link_targets",
+    "partition_delete_links",
+    "resolve_work_item_types",
+]
