+++ src/mcp_server_polarion/tools/_shared/guard/test_runs.py
@@ -1,0 +1,208 @@
+"""Test-run write guards: enums (``testing`` context), template resolution,
+custom-field keys.
+"""
+
+from __future__ import annotations
+
+from collections.abc import Iterable
+
+from mcp_server_polarion.core.client import PolarionClient
+from mcp_server_polarion.core.exceptions import (
+    PolarionAuthError,
+    PolarionError,
+    PolarionNotFoundError,
+)
+from mcp_server_polarion.tools._shared.cache import (
+    get_test_run_custom_keys,
+    invalidate_test_run_custom_keys,
+    store_test_run_custom_keys,
+)
+from mcp_server_polarion.tools._shared.custom_fields import (
+    STANDARD_TEST_RUN_ATTRIBUTES,
+)
+from mcp_server_polarion.tools._shared.guard._common import (
+    _GUARD_PAGE_SIZE,
+    _custom_keys_from_data_list,
+    _reject_unknown_custom_keys,
+)
+from mcp_server_polarion.tools._shared.guard._errors import (
+    _unauthorized_write_block,
+    _unreachable_write_block,
+)
+from mcp_server_polarion.tools._shared.guard.enums import _check_project_enum_roles
+from mcp_server_polarion.tools._shared.helpers import (
+    encode_path_segment,
+    format_option_list,
+)
+
+
+async def guard_test_run_enums(
+    client: PolarionClient,
+    project_id: str,
+    *,
+    type: str | None = None,
+    status: str | None = None,
+) -> None:
+    """Validate test-run ``type``/``status`` against the ``testing``-context
+    project enumerations — testruns have no ``getAvailableOptions`` endpoint,
+    and the ``~`` wildcard context does not resolve these enums.
+    """
+    await _check_project_enum_roles(
+        client,
+        project_id,
+        "testrun-type",
+        [type] if type else [],
+        field_label="test run type",
+        discovery_hint="use list_test_runs to see values in use.",
+        context="testing",
+    )
+    await _check_project_enum_roles(
+        client,
+        project_id,
+        "testrun-status",
+        [status] if status else [],
+        field_label="test run status",
+        discovery_hint="use list_test_runs to see values in use.",
+        context="testing",
+    )
+
+
+async def guard_test_run_templates(
+    client: PolarionClient,
+    project_id: str,
+    template_ids: Iterable[str],
+) -> None:
+    """Resolve each template id before the write — Polarion doesn't validate
+    relationship targets. Run instances are rejected too: ``isTemplate`` is
+    served only (as ``true``) on templates, absent on instances.
+    """
+    for template_id in sorted({t for t in template_ids if t}):
+        path = (
+            f"/projects/{encode_path_segment(project_id)}"
+            f"/testruns/{encode_path_segment(template_id)}"
+        )
+        try:
+            response = await client.get(
+                path, params={"fields[testruns]": "id,isTemplate"}
+            )
+        except PolarionNotFoundError as exc:
+            raise ValueError(
+                f"Test run template '{template_id}' not found in project "
+                f"'{project_id}'. Use list_test_runs(templates=True) to "
+                f"discover template ids."
+            ) from exc
+        except PolarionAuthError as exc:
+            raise _unauthorized_write_block("test run templates", project_id) from exc
+        except PolarionError as exc:
+            raise _unreachable_write_block(
+                "test run templates", project_id, exc
+            ) from exc
+
+        data = response.get("data", {})
+        attributes = data.get("attributes") if isinstance(data, dict) else None
+        is_template = (
+            attributes.get("isTemplate") if isinstance(attributes, dict) else None
+        )
+        if is_template is not True:
+            raise ValueError(
+                f"Test run '{template_id}' in project '{project_id}' is a run "
+                f"instance, not a template. Use list_test_runs(templates=True) "
+                f"to discover template ids."
+            )
+
+
+async def _fetch_test_run_custom_keys(
+    client: PolarionClient,
+    project_id: str,
+) -> frozenset[str]:
+    """Union of custom-field keys sampled from the project's runs and
+    templates — testrun custom fields are project config (no type axis, no
+    SQL needed). Cached even if empty.
+    """
+    path = f"/projects/{encode_path_segment(project_id)}/testruns"
+    keys: set[str] = set()
+    for templates in (False, True):
+        base_params: dict[str, str | int] = {
+            "fields[testruns]": "@all",
+            "page[size]": _GUARD_PAGE_SIZE,
+        }
+        if templates:
+            base_params["templates"] = "true"
+        page_number = 1
+        while True:
+            try:
+                response = await client.get(
+                    path, params={**base_params, "page[number]": page_number}
+                )
+            except PolarionAuthError as exc:
+                raise _unauthorized_write_block(
+                    "custom_fields keys", project_id
+                ) from exc
+            except PolarionError as exc:
+                raise _unreachable_write_block(
+                    "custom_fields keys", project_id, exc
+                ) from exc
+            data = response.get("data", [])
+            if not isinstance(data, list):
+                break
+            keys.update(
+                _custom_keys_from_data_list(response, STANDARD_TEST_RUN_ATTRIBUTES)
+            )
+            if len(data) < _GUARD_PAGE_SIZE:
+                break
+            page_number += 1
+
+    result = frozenset(keys)
+    store_test_run_custom_keys(project_id, result)
+    return result
+
+
+async def _check_test_run_custom_keys(
+    client: PolarionClient,
+    project_id: str,
+    custom_fields: dict[str, object],
+) -> None:
+    """Test-run mirror of :func:`_check_work_item_custom_keys` (project scope)."""
+    schema = get_test_run_custom_keys(project_id)
+    fetched_fresh = schema is None
+    if schema is None:
+        schema = await _fetch_test_run_custom_keys(client, project_id)
+
+    if all(key in schema for key in custom_fields):
+        return
+
+    # Unknown key may be admin-added since caching; refetch once before rejecting.
+    if not fetched_fresh:
+        invalidate_test_run_custom_keys(project_id)
+        schema = await _fetch_test_run_custom_keys(client, project_id)
+
+    if not schema:
+        raise RuntimeError(
+            f"Cannot verify custom_fields {format_option_list(custom_fields)} for "
+            f"test runs in project '{project_id}': no existing test run has custom "
+            f"fields populated, so the schema can't be sampled. Refusing the write "
+            f"-- an unknown key ghosts silently (invisible to UI/Lucene). Do not "
+            f"create runs to work around this; ask the user to confirm these "
+            f"custom-field ids exist for test runs."
+        )
+
+    _reject_unknown_custom_keys(
+        custom_fields,
+        schema,
+        scope=f"test runs in project '{project_id}'",
+        discovery_tool="sample of existing runs",
+    )
+
+
+async def guard_test_run_custom_fields(
+    client: PolarionClient,
+    project_id: str,
+    custom_fields: dict[str, object],
+) -> None:
+    """Validate ``custom_fields`` keys before a test-run write. Keys only —
+    testruns have no ``getAvailableOptions``, so enum-typed *values* defer to
+    Polarion (wrong option ids there ghost; keys are the guardable axis).
+    """
+    if not custom_fields:
+        return
+    await _check_test_run_custom_keys(client, project_id, custom_fields)
